Move teams stuff to module data.teams
